Move utility classes into namespace to ...
avoid collisions with Nake.Utility exposed classes in tests
