+++ Source/Nake/Program.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
+
+using Nake.Utility;
+
 using static System.Environment;
 
 using AsyncTask = System.Threading.Tasks.Task;
